Introduce doclets to transform javadocs into TSDocs (#81)
As part of the goals of this tool we are trying to also have
documentations for the TypeScript definition generated from the Java
JsTypes, In the original implementation using annotation processors
didnt help us in doing so, the annotation processor API is limited when
it comes to read and parse the JavaDocs from the Java source elements,
for that reason we went initially with a simple solution by copying the
JavaDocs String from the JavaElements as is and put it on the generated
TypeScript members.

The problem with the original solution is that it seemed to be ok when
reading it directly from the source code, but when using any TypeScript
doc tool to generate some actual html docs for example we end up with
many docs that do not follow the TypeScript docs conventions and style,
links from JavaDocs wont match the links from TypeScript and so `@code`,
and some times we end with a broken HTML tags.

The proposed solution by this pull request is to use Java doc tools to
actually generate the TypeScript and have a better API to actually parse
the JavaDocs and transform them into a TypeScript docs, and doclets can
do a better job here, we are keeping the processor and refactoring the
code to actually share the APIs that generated the TypeScript
definitions between the annotation processor and the doclet while
letting each handle the docs on its own, so for now the doclet will
generate better TypeScript docs than the processor, and will correctly
transform links and code tags from JavaDocs.

Fix #80

(cherry picked from commit d9d60164f88117b43e978339d85b003b4fe7cc86)